// src/routes/interface.ts
export interface Customer{  
    customerid: number;
    fullname: string; 
    email: string
    gender: string; 
    phone: string; 
    address: string; 
    notes: string;  
}

export interface User{
    id: number;
    username: string;
    email: string;
    password: string
}

// export default Customer;